+++ back-end/src/main/java/com/tdd/backend/user/controller/UserController.java
@@ -4,20 +4,15 @@ import javax.validation.Valid;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestHeader;
 import org.springframework.web.bind.annotation.RestController;
 
-import com.tdd.backend.auth.LoginUser;
 import com.tdd.backend.auth.data.JwtTokenPairResponse;
-import com.tdd.backend.auth.jwt.RefreshTokenStorage;
 import com.tdd.backend.user.data.UserCreate;
 import com.tdd.backend.user.data.UserLogin;
-import com.tdd.backend.user.data.UserToken;
 import com.tdd.backend.user.exception.DuplicateEmailException;
 import com.tdd.backend.user.service.UserService;
 
@@ -40,6 +35,7 @@ public class UserController {
 		return new ResponseEntity<>(HttpStatus.OK);
 	}
 
+	@Operation(summary = "유저 이메일 중복여부 체크", description = "디비에 이미 저장되어 잇을 시 DuplicateEmailException 응답")
 	@GetMapping("/users/validation/{email}")
 	public void userEmailValidCheck(@PathVariable String email) {
 		if (userService.isDuplicateEmail(email)) {
@@ -51,27 +47,5 @@ public class UserController {
 	@PostMapping("/login")
 	public ResponseEntity<JwtTokenPairResponse> login(@RequestBody @Valid UserLogin userLogin) {
 		return ResponseEntity.ok(userService.login(userLogin));
-	}
-
-	//TODO : 역할에 따라 토큰 서비스 클래스 분리
-
-	// RTK이며, 요청이 POST /reissue인 경우 재발급을 진행한다.
-	@PostMapping("/reissue")
-	public ResponseEntity<JwtTokenPairResponse> refreshAccessToken(
-		@RequestHeader("Authorization") String refreshToken) {
-		log.info("> refresh token : {}", refreshToken);
-		return ResponseEntity.ok(userService.reIssueToken(refreshToken));
-	}
-
-	//사용자가 로그아웃을 하면 저장소에서 Refresh Token을 삭제하여 사용이 불가능하도록 한다.
-	//todo : redirect??
-	@DeleteMapping("/logout")
-	public void logout(@RequestHeader("Authorization") String refreshToken) {
-		RefreshTokenStorage.deleteCache(refreshToken);
-	}
-
-	@GetMapping("/test/auth")
-	public String testAuth(@LoginUser UserToken userToken) {
-		return "JWT IS AWESOME " + userToken.getId();
 	}
 }
